backend/pyproject.toml
[project]
name = "english-cafe-backend"
version = "1.0.0"
description = "英会話カフェWebサイトのバックエンドAPI"
authors = [
    {name = "English Cafe Team"}
]
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.32.1",
    "python-multipart==0.0.20",
    "sqlalchemy==2.0.36",
    "alembic==1.14.0",
    "asyncpg==0.30.0",
    "psycopg2-binary==2.9.10",
    "pydantic==2.10.3",
    "pydantic-settings==2.7.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "slowapi==0.1.9",
    "aiosmtplib==3.0.2",
    "jinja2==3.1.4",
    "email-validator>=2.3.0",
]

[dependency-groups]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "pytest-cov==6.0.0",
    "httpx==0.28.1",
    "ruff==0.8.4",
    "safety==3.2.11",
    "aiosqlite>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 88
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "PIE", "T20"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--tb=short",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=xml",
]
